Extract DIMPLE class + addgene into DIMPLE/core.py
Phase 0 of the module-migration refactor. The DIMPLE class and addgene
FASTA loader move verbatim into a new leaf module, DIMPLE/core.py, which
depends only on DIMPLE.utilities and biopython. DIMPLE.py imports and
re-exports both for back-compat; run_settings.py imports the class from
core directly.

This breaks the import cycle that would otherwise arise once the
functional modules (primers, qc, fragment_layout, oligo_assembly,
outputs) own their implementations: core is now a true leaf, functional
modules depend on it, and the DIMPLE.py orchestrator depends on both.

No behavior change; full suite green (63 passed).
